// src/app/generated/graphql.ts
export type Maybe<T> = T | null;

export interface ClientInput {
	name?: Maybe<string>;

	company?: Maybe<string>;

	street?: Maybe<string>;

	city?: Maybe<string>;

	psc?: Maybe<string>;

	country?: Maybe<string>;

	ico?: Maybe<string>;

	dic?: Maybe<string>;

	icdph?: Maybe<string>;

	post_company?: Maybe<string>;

	post_street?: Maybe<string>;

	post_city?: Maybe<string>;

	post_psc?: Maybe<string>;

	post_country?: Maybe<string>;

	emails?: Maybe<(Maybe<EmailInputs>)[]>;

	phones?: Maybe<(Maybe<PhoneInput>)[]>;

	price_hour?: Maybe<number>;

	contact_person?: Maybe<string>;

	description?: Maybe<string>;
}

export interface EmailInputs {
	email?: Maybe<string>;

	comment?: Maybe<string>;
}

export interface PhoneInput {
	phone?: Maybe<string>;

	comment?: Maybe<string>;
}
// ...
export type DateTime = any;
// ...
export type AddClientVariables = {
	input: ClientInput;
};

export type AddClientMutation = {
	__typename?: "Mutation";

	addClient: AddClientAddClient;
};

export type AddClientAddClient = {
	__typename?: "ClientResponseMutation";

	ok: boolean;

	message: string;

	client: Maybe<AddClientClient>;
};
// ...
export type ClientPartsFragment = {
	__typename?: "Client";

	id: Maybe<string>;

	name: Maybe<string>;
};

// ====================================================
// Scalars
// ====================================================

// ====================================================
// Interfaces
// ====================================================

export interface ClientInterface {
	name?: Maybe<string>;

	company?: Maybe<string>;

	street?: Maybe<string>;

	city?: Maybe<string>;

	psc?: Maybe<number>;

	country?: Maybe<string>;

	ico?: Maybe<number>;

	dic?: Maybe<number>;

	icdph?: Maybe<number>;

	post_company?: Maybe<string>;

	post_street?: Maybe<string>;

	post_city?: Maybe<string>;

	post_psc?: Maybe<number>;

	post_country?: Maybe<string>;

	emails?: Maybe<(Maybe<ClientEmails>)[]>;

	phones?: Maybe<(Maybe<ClientPhones>)[]>;

	price_hour?: Maybe<number>;

	contact_person?: Maybe<string>;

	description?: Maybe<string>;

	created_at?: Maybe<DateTime>;

	deleted_at?: Maybe<DateTime>;
}

export interface ResponseMutation {
	code?: Maybe<string>;

	ok: boolean;

	message: string;
}

// ====================================================
// Types
// ====================================================

export interface Query {
	_empty?: Maybe<string>;

	client: Client;

	clients: (Maybe<Client>)[];

	task?: Maybe<Task>;

	tasks?: Maybe<(Maybe<Task>)[]>;

	user: User;

	users: (Maybe<User>)[];
}

export interface Client extends ClientInterface {
	id?: Maybe<string>;

	name?: Maybe<string>;

	company?: Maybe<string>;

	street?: Maybe<string>;

	city?: Maybe<string>;

	psc?: Maybe<number>;

	country?: Maybe<string>;

	ico?: Maybe<number>;

	dic?: Maybe<number>;

	icdph?: Maybe<number>;

	post_company?: Maybe<string>;

	post_street?: Maybe<string>;

	post_city?: Maybe<string>;

	post_psc?: Maybe<number>;

	post_country?: Maybe<string>;

	emails?: Maybe<(Maybe<ClientEmails>)[]>;

	phones?: Maybe<(Maybe<ClientPhones>)[]>;

	price_hour?: Maybe<number>;

	contact_person?: Maybe<string>;

	description?: Maybe<string>;

	created_at?: Maybe<DateTime>;

	deleted_at?: Maybe<DateTime>;

	tasks?: Maybe<(Maybe<Task>)[]>;
}

export interface ClientEmails {
	id?: Maybe<string>;

	email?: Maybe<string>;

	comment?: Maybe<string>;
}

export interface ClientPhones {
	id?: Maybe<string>;

	phone?: Maybe<string>;

	comment?: Maybe<string>;
}

export interface Task {
	id: string;

	name: string;

	description?: Maybe<string>;

	completed?: Maybe<boolean>;

	archive?: Maybe<boolean>;

	created_at?: Maybe<DateTime>;

	deleted_at?: Maybe<DateTime>;

	client?: Maybe<Client>;
}

export interface User {
	id?: Maybe<string>;

	name?: Maybe<string>;

	email?: Maybe<string>;

	password?: Maybe<string>;
}

export interface Mutation {
	_empty?: Maybe<string>;

	addClient: ClientResponseMutation;

	updateClient: ClientResponseMutation;

	deleteClient: ClientResponseMutation;

	addTask: TaskResponseMutation;

	updateTask: TaskResponseMutation;

	deleteTask: TaskResponseMutation;

	addUser: UserResponseMutation;

	updateUser: UserResponseMutation;

	deleteUser: UserResponseMutation;
}

export interface ClientResponseMutation extends ResponseMutation {
	code?: Maybe<string>;

	ok: boolean;

	message: string;

	client?: Maybe<Client>;
}

export interface TaskResponseMutation extends ResponseMutation {
	code?: Maybe<string>;

	ok: boolean;

	message: string;

	task?: Maybe<Task>;
}

export interface UserResponseMutation extends ResponseMutation {
	code?: Maybe<string>;

	ok: boolean;

	message: string;

	user?: Maybe<User>;
}
// ...
export const ClientPartsFragment = gql`
	fragment ClientParts on Client {
		id
		name
	}
`;

// ====================================================
// Apollo Services
// ====================================================

@Injectable({
	providedIn: "root"
})
export class AddClientGQL extends Apollo.Mutation<
	AddClientMutation,
	AddClientVariables
> {
	document: any = gql`
		mutation AddClient($input: ClientInput!) {
			addClient(input: $input) {
				ok
				message
				client {
					...ClientParts
				}
			}
		}

		${ClientPartsFragment}
	`;
}